Adds timezone support
This timezone is the timezone you want the dates to be written in the
database. Local dates are converted to the timezone you want and then
written. Reading will do the oposite. Without this timezone setting,
dates will be written in UTC (default timezone = "Z").

There's also a tweak on the typeCasting, not finished yet, that enables
people to make a custom typeCast function and use the default as
fallback.