fix: PromiseConnection missing interface functions from Connection
`PromiseConnection` was missing some functions `Connection` implements.
Hence it was unable to use the `Promise` wrapped connection like
the "real" connection.

This commit adds implementation for all missing asynchronous functions.
Additionally a whitelist is used to determine, which synchronous
functions are implementd by a facade function, passing the call to
the internally used connection instance.