Make Promise Pool wrapper extend EventEmitter
Fix #468

Introduce a new PromisePool constructor into promise.js. PromisePool
inherits from EventEmitter, and the following events are re-emitted
from the underlying Pool instance to it:
 - acquire
 - connection
 - enqueue
 - release

Methods like `on` can't just be bound to the Pool as `this` would be
wrong in the callback: e.g. `this.query` wouldn't return a Promise.