tests: decouple fake server from MYSQL_HOST
Fake servers are being created on the host specified by the MYSQL_HOST
environment variable. This can lead to issues when the real MySQL server
instance used for the integration tests is not running on the same host,
for instance, in Docker-based CI environments.

This patch ensures fake servers are explicitely created in the
host/container where the test suite is running.