Interim fix for using prepared statements with MySQL 8.0.22 and above.
When running execute(), any placeholder parameter value with the JavaScript type 'number' will now be sent to MySQL via the binary protocol as a MySQL string instead of as a MySQL double.

Resolves issue #1239, where it was reported that prepared statements that are called with parameters having the JavaSCript type 'number' are failing with the error "Incorrect arguments to mysqld_stmt_execute".

It was determined that this was due to a change made to prepared statement handling in MySQL 8.0.22 and that this could be overcome by converting numerical parameter values to strings.

The long-term fix requires the implementation of data type conversion of parameters using the type hints provided by MySQL in the COM_STMT_PREPARE Response.

For more information:

https://github.com/sidorares/node-mysql2/issues/1239
https://dev.mysql.com/worklog/task/?id=9384