handle connection close error message
A custom ERR_PACKET is sent by the server on MySQL 8.0.24 (and higher)
when the server disconnects inactive clients.

This causes some integration tests that destroy the socket connection to
fail due to packet number and order mismatch.

The patch introduces logic to handle this potential new and unexpected
packet and report the custom server message to the application.