fix non-standard invalid date test
NO_ZERO_DATE mode and NO_ZERO_IN_DATE mode are part of the strict SQL
mode used by MySQL 8.0. Non-standard dates like '0000-00-00' are now
stored and retrieved as NULL.

This patch introduces changes to one of the integration tests to ensure
this change in behavior is accounted for with the latest MySQL server
versions.