avoid column length expectations for integers
The column length returned as result set metadata from the server only
makes sense for DECIMAL column data types. For any other type, the value
is non-deterministic and this kind of display width for integer data
types was deprecated on MySQL 8.0.17.

https://dev.mysql.com/doc/refman/8.0/en/numeric-type-syntax.html

This patch updates some integration tests that are failing due to an
incorrect assumption about the "columnLength" value, by avoiding to assert
any expectation about it altogether. In this case, the tests are expecting
a statement to yield a column length of a number of digits, which is the
case on MySQL 8.0.24 or older versions. However, MySQL 8.0.25 reports an
additional digit to accommodate the sign for signed integers.

The breaking change was introduced by the server to address the issues
described in the following bug report:

https://bugs.mysql.com/bug.php?id=99567