travis: save docker images on local cache
The recently introduced rate limits on Docker Hub are constantly being
reached by every TravisCI because each individual job that creates a
MySQL container, tries to pull the base image from Docker Hub. With a
not-so-small test matrix, this issue surfaces pretty fast.

This patch introduces a local cache to save docker images and be able to
re-use them on different jobs and avoid constantly pulling them from
Docker Hub.