add handshake fatal error handling
Errors that occur during the handshake stage should be fatal and should
cause the connection to be closed.

Currently, although unexpected errors such as ones created by
unexpected packets sent by the server, are being handled, they are not
considered fatal and do not result in closing the connection. On the other
hand, errors reported by client-side authentication plugins are being
ignored and are not reported to the application.

Since errors reported by server-side authentication plugins are
considered fatal and cause the connection to be closed, the fact that
the same does not happen for errors generated on the client-side leads
to some inconsistencies and forces custom authentication plugin
implementations to forcibly close the connection on their own.

This patch ensures that both this kind of errors are considered fatal and
cause the connection to be closed.