Optimize Query.row call
Query.row was called with two arguments but the declaration has only one
argument. By changing it to two-argument function, it eliminates the
need for V8 to create an additional Argument Adapter Trampoline to
bridge between the caller and the callee. This optimization is mainly
for Node JS < 16.0.0 as V8 8.9 has already incorporated some
optimization for this case.

https://v8.dev/blog/adaptor-frame