add support for multi-factor authentication
MySQL server 8.0.28 introduced support for multi-factor authentication.
User accounts can now be configured with up to three authentication
plugins and corresponding passwords or additional options. Although not
all authentication plugins can be used for the 2nd and 3rd factor, a
MySQL client should be flexible enough to accommodate those that can.

This patch extends the client handshake workflow to account for an
additional "AuthNextFactor" packet that is sent by the server after a
successful authentication factor. When the user account is configured
with more than one authentication factor, this will happen, at most,
after the 1st factor and additionally after the 2nd (if it is not the
last).

The packet exchange for each factor remains unchanged, and after the
last authentication factor, the server either replies with an OK_Packet
or an ERR_Packet depending on whether or not the access is granted,
similar to what already happens in a single-factor authentication setup.